+++ src/main/java/io/github/slimefunguguproject/bump/implementation/items/weapons/SkySword.java
@@ -10,7 +10,7 @@ import org.bukkit.util.Vector;
 import io.github.slimefunguguproject.bump.core.services.sounds.BumpSound;
 import io.github.slimefunguguproject.bump.implementation.Bump;
 import io.github.slimefunguguproject.bump.implementation.BumpItems;
-import io.github.slimefunguguproject.bump.utils.Keys;
+import io.github.slimefunguguproject.bump.utils.constant.Keys;
 import io.github.thebusybiscuit.slimefun4.api.recipes.RecipeType;
 import io.github.thebusybiscuit.slimefun4.implementation.SlimefunItems;
 import io.github.thebusybiscuit.slimefun4.libraries.dough.data.persistent.PersistentDataAPI;
