+++ src/main/java/io/github/slimefunguguproject/bump/implementation/listeners/SkySwordListener.java
@@ -9,7 +9,7 @@ import org.bukkit.event.entity.EntityDamageEvent;
 
 import io.github.slimefunguguproject.bump.implementation.Bump;
 import io.github.slimefunguguproject.bump.implementation.items.weapons.SkySword;
-import io.github.slimefunguguproject.bump.utils.Keys;
+import io.github.slimefunguguproject.bump.utils.constant.Keys;
 import io.github.thebusybiscuit.slimefun4.libraries.dough.data.persistent.PersistentDataAPI;
 
 /**
