+++ src/main/java/io/github/slimefunguguproject/bump/implementation/tasks/WeaponProjectileTask.java
@@ -11,7 +11,7 @@ import com.google.common.base.Preconditions;
 import org.bukkit.entity.Projectile;
 
 import io.github.slimefunguguproject.bump.implementation.Bump;
-import io.github.slimefunguguproject.bump.utils.Keys;
+import io.github.slimefunguguproject.bump.utils.constant.Keys;
 import io.github.thebusybiscuit.slimefun4.implementation.Slimefun;
 import io.github.thebusybiscuit.slimefun4.libraries.dough.data.persistent.PersistentDataAPI;
